feat: update mirror demo
